feat(shape): Finished MLOD Reading
