+++ src/File Formats/BisUtils.P3D/Extensions/RvShapeExtensions.cs
@@ -1,17 +1,18 @@
 ﻿namespace BisUtils.P3D.Extensions;
 
+using Models;
 using Models.Data;
 using Models.Lod;
 
-public static class RvLodExtensions
+public static class RvShapeExtensions
 {
-    public static IRVShapeData? LocateLevel(this IRVLod lod, float resolution, out int foundLevelPosition)
+    public static IRVShapeData? LocateLevel(this IRVShape lod, float resolution, out int foundLevelPosition)
     {
         var minDifference = Math.Abs(resolution) * 0.00001f;
         IRVShapeData? found = null;
         foundLevelPosition = -1;
         var levelPosition = -1;
-        foreach (var level in lod.LodLevels)
+        foreach (var level in lod.LevelsOfDetail)
         {
             levelPosition++;
             var difference = Math.Abs(resolution - level.Resolution.Value);
